backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg[binary]==3.2.13
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<5.0.0  # passlib 1.7.4 requires bcrypt < 5.0.0
python-multipart==0.0.6
paddlepaddle==3.2.2
paddleocr==2.10.0
spacy==3.8.11
transformers==4.35.2
torch>=2.5.0
openai==1.3.7
boto3==1.29.7
minio==7.2.0
pillow==10.1.0
faker==20.1.0
reportlab==4.0.7
python-dotenv==1.0.0
httpx==0.25.2
pdf2image==1.16.3
email-validator>=2.0.0

